requirements.txt
Django
djangorestframework
django-cors-headers
celery
django-celery-beat
django-cleanup
Pillow
redis